# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API document of Intelligent Terminal Backend",
        "title": "Intelligent Terminal Backend API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
...
        "/user/avatarUpdate": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to upload image and replace user's avatar as the uploaded image",
                "parameters": [
                    {
                        "type": "file",
                        "description": "the avatar image file selected by the user",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {}
            }
        },
        "/user/avatarUpdateBase64": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to update the avatar in the form of base64 string",
                "parameters": [
                    {
                        "description": "base64 string",
                        "name": "\"base64\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.AvatarUpdateBase64Req"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/basicInfo": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "used to get user's basic info",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to get the target user's detailed information",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/follow": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to follow other people",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/followerList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to get the user's follower list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "used to authorize user and return jwt token, id",
                "parameters": [
                    {
                        "description": "the passed-in parameter of login function",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserLoginReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/modify": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to modify the user's personal information",
                "parameters": [
                    {
                        "description": "the passed-in parameter of modify function",
                        "name": "UserModifyReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserModifyReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/momentList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to show the moment list of specified user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/user/passwordChange": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to modify the user's password",
                "parameters": [
                    {
                        "description": "old password and new password",
                        "name": "UserPasswordChangeReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserPasswordChangeReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "used to register new account",
                "parameters": [
                    {
                        "description": "the passed-in parameter of register function",
                        "name": "UserRegisterReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserRegisterReq"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/user/subscribedList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User"
                ],
                "summary": "used to get the user's subscribed list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "request.AvatarUpdateBase64Req": {
            "type": "object",
            "properties": {
                "base64": {
                    "type": "string"
                }
            }
        },
...
        "request.UserLoginReq": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "request.UserModifyReq": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.UserPasswordChangeReq": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "request.UserRegisterReq": {
            "type": "object",
            "properties": {
                "account": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}